+++ cpp/arcticdb/CMakeLists.txt
@@ -187,7 +187,6 @@ set(arcticdb_srcs
         entity/frame_and_descriptor.hpp
         entity/index_range.hpp
         entity/key.hpp
-        entity/key-inl.hpp
         entity/metrics.hpp
         entity/data_error.hpp
         entity/native_tensor.hpp
@@ -246,9 +245,7 @@ set(arcticdb_srcs
         storage/library_index.hpp
         storage/library_manager.hpp
         storage/azure/azure_storage.hpp
-        storage/azure/azure_storage-inl.hpp
         storage/lmdb/lmdb_storage.hpp
-        storage/lmdb/lmdb_storage-inl.hpp
         storage/memory/memory_storage.hpp
         storage/memory/memory_storage.cpp
         storage/mongo/mongo_client.hpp
@@ -256,11 +253,8 @@ set(arcticdb_srcs
         storage/mongo/mongo_storage.hpp
         storage/object_store_utils.hpp
         storage/s3/nfs_backed_storage.hpp
-        storage/s3/nfs_backed_storage-inl.hpp
         storage/s3/s3_client_accessor.hpp
-        storage/s3/s3_storage-inl.hpp
         storage/s3/s3_storage_tool.hpp
-        storage/s3/s3_storage_tool-inl.hpp
         storage/storage_factory.hpp
         storage/storage_options.hpp
         storage/storage.hpp
